(refactor) resources
- Add objects and additional utils
- Refactor folder structure
