--- app/src/main/java/com/example/sam/moviebox/repository/resources/Status.java
@@ -1,7 +1,0 @@
-package com.example.sam.moviebox.repository.resources;
-
-public enum Status {
-    SUCCESS,
-    ERROR,
-    LOADING
-}
