+++ app/src/main/java/com/example/sam/moviebox/repository/resources/data/LiveDataCallAdapterFactory.java
@@ -1,4 +1,4 @@
-package com.example.sam.moviebox.repository.util;
+package com.example.sam.moviebox.repository.resources.data;
 
 import android.arch.lifecycle.LiveData;
