+++ app/src/main/java/com/example/sam/moviebox/repository/resources/data/Objects.java
@@ -1,0 +1,13 @@
+package com.example.sam.moviebox.repository.resources.data;
+
+public class Objects {
+    public static boolean equals(Object o1, Object o2){
+        if (o1 == null){
+            return o2 == null;
+        }
+        if (o2 == null){
+            return false;
+        }
+        return o1.equals(o2);
+    }
+}
