+++ app/src/main/java/com/example/sam/moviebox/repository/resources/utils/Status.java
@@ -1,0 +1,7 @@
+package com.example.sam.moviebox.repository.resources.utils;
+
+public enum Status {
+    SUCCESS,
+    ERROR,
+    LOADING
+}
